@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "coveralls": "cat ./reports/coverage/lcov.info | coveralls",
-    "lint": "eslint ./",
+    "lint": "eslint ./lib",
     "test": "npm run lint && npm run unit-test",
     "unit-test": "istanbul cover --dir reports/coverage node_modules/mocha/bin/_mocha tests/**/*.js -- --reporter dot"
   },
@@ -23,9 +23,11 @@
   "homepage": "https://github.com/intellicode/eslint-plugin-react-native",
   "bugs": "https://github.com/intellicode/eslint-plugin-react-native/issues",
   "devDependencies": {
-    "babel-eslint": "4.1.8",
+    "babel-eslint": "6.0.0",
     "coveralls": "2.11.4",
-    "eslint": "2.0.0",
+    "eslint": "2.4.0",
+    "eslint-config-airbnb": "6.2.0",
+    "eslint-plugin-react": "4.2.3",
     "istanbul": "0.4.0",
     "mocha": "2.3.3"
   },
