@@ -4,7 +4,7 @@
  */
 'use strict';
 
-var util = require('util');
+const util = require('util');
 
 /**
  * Components
@@ -12,7 +12,7 @@ var util = require('util');
  */
 function Components() {
   this._list = {};
-  this._getId = function(node) {
+  this._getId = function (node) {
     return node && node.range.join(':');
   };
 }
@@ -23,8 +23,8 @@ function Components() {
  * @param {ASTNode} node The AST node being added.
  * @param {Number} confidence Confidence in the component detection (0=banned, 1=maybe, 2=yes)
  */
-Components.prototype.add = function(node, confidence) {
-  var id = this._getId(node);
+Components.prototype.add = function (node, confidence) {
+  const id = this._getId(node);
   if (this._list[id]) {
     if (confidence === 0 || this._list[id].confidence === 0) {
       this._list[id].confidence = 0;
@@ -35,7 +35,7 @@ Components.prototype.add = function(node, confidence) {
   }
   this._list[id] = {
     node: node,
-    confidence: confidence
+    confidence: confidence,
   };
 };
 
@@ -45,8 +45,8 @@ Components.prototype.add = function(node, confidence) {
  * @param {ASTNode} node The AST node being searched.
  * @returns {Object} Component object, undefined if the component is not found
  */
-Components.prototype.get = function(node) {
-  var id = this._getId(node);
+Components.prototype.get = function (node) {
+  const id = this._getId(node);
   return this._list[id];
 };
 
@@ -56,14 +56,15 @@ Components.prototype.get = function(node) {
  * @param {ASTNode} node The AST node being updated.
  * @param {Object} props Additional properties to add to the component.
  */
-Components.prototype.set = function(node, props) {
-  while (node && !this._list[this._getId(node)]) {
-    node = node.parent;
+Components.prototype.set = function (node, props) {
+  let currentNode = node;
+  while (currentNode && !this._list[this._getId(currentNode)]) {
+    currentNode = node.parent;
   }
-  if (!node) {
+  if (!currentNode) {
     return;
   }
-  var id = this._getId(node);
+  const id = this._getId(currentNode);
   this._list[id] = util._extend(this._list[id], props);
 };
 
@@ -73,9 +74,9 @@ Components.prototype.set = function(node, props) {
  *
  * @returns {Object} Components list
  */
-Components.prototype.list = function() {
-  var list = {};
-  for (var i in this._list) {
+Components.prototype.list = function () {
+  const list = {};
+  for (const i in this._list) {
     if (!this._list.hasOwnProperty(i) || this._list[i].confidence < 2) {
       continue;
     }
@@ -90,9 +91,9 @@ Components.prototype.list = function() {
  *
  * @returns {Number} Components list length
  */
-Components.prototype.length = function() {
-  var length = 0;
-  for (var i in this._list) {
+Components.prototype.length = function () {
+  let length = 0;
+  for (const i in this._list) {
     if (!this._list.hasOwnProperty(i) || this._list[i].confidence < 2) {
       continue;
     }
@@ -102,12 +103,11 @@ Components.prototype.length = function() {
 };
 
 function componentRule(rule, context) {
-
-  var sourceCode = context.getSourceCode();
-  var components = new Components();
+  const sourceCode = context.getSourceCode();
+  const components = new Components();
 
   // Utilities for component detection
-  var utils = {
+  const utils = {
 
     /**
      * Check if the node is a React ES5 component
@@ -115,7 +115,7 @@ function componentRule(rule, context) {
      * @param {ASTNode} node The AST node being checked.
      * @returns {Boolean} True if the node is a React ES5 component, false if not
      */
-    isES5Component: function(node) {
+    isES5Component: function (node) {
       if (!node.parent) {
         return false;
       }
@@ -128,7 +128,7 @@ function componentRule(rule, context) {
      * @param {ASTNode} node The AST node being checked.
      * @returns {Boolean} True if the node is a React ES6 component, false if not
      */
-    isES6Component: function(node) {
+    isES6Component: function (node) {
       if (!node.superClass) {
         return false;
       }
@@ -141,8 +141,8 @@ function componentRule(rule, context) {
      * @param {ASTNode} node The AST node being checked (must be a ReturnStatement).
      * @returns {Boolean} True if the node is returning JSX, false if not
      */
-    isReturningJSX: function(node) {
-      var property;
+    isReturningJSX: function (node) {
+      let property;
       switch (node.type) {
         case 'ReturnStatement':
           property = 'argument';
@@ -154,11 +154,11 @@ function componentRule(rule, context) {
           return false;
       }
 
-      var returnsJSX =
+      const returnsJSX =
         node[property] &&
         node[property].type === 'JSXElement'
       ;
-      var returnsReactCreateElement =
+      const returnsReactCreateElement =
         node[property] &&
         node[property].callee &&
         node[property].callee.property &&
@@ -173,7 +173,7 @@ function componentRule(rule, context) {
      *
      * @returns {ASTNode} component node, null if we are not in a component
      */
-    getParentComponent: function() {
+    getParentComponent: function () {
       return (
         utils.getParentES6Component() ||
         utils.getParentES5Component() ||
@@ -186,10 +186,10 @@ function componentRule(rule, context) {
      *
      * @returns {ASTNode} component node, null if we are not in a component
      */
-    getParentES5Component: function() {
-      var scope = context.getScope();
+    getParentES5Component: function () {
+      let scope = context.getScope();
       while (scope) {
-        var node = scope.block && scope.block.parent && scope.block.parent.parent;
+        const node = scope.block && scope.block.parent && scope.block.parent.parent;
         if (node && utils.isES5Component(node)) {
           return node;
         }
@@ -203,12 +203,12 @@ function componentRule(rule, context) {
      *
      * @returns {ASTNode} component node, null if we are not in a component
      */
-    getParentES6Component: function() {
-      var scope = context.getScope();
+    getParentES6Component: function () {
+      let scope = context.getScope();
       while (scope && scope.type !== 'class') {
         scope = scope.upper;
       }
-      var node = scope && scope.block;
+      const node = scope && scope.block;
       if (!node || !utils.isES6Component(node)) {
         return null;
       }
@@ -220,13 +220,16 @@ function componentRule(rule, context) {
      *
      * @returns {ASTNode} component node, null if we are not in a component
      */
-    getParentStatelessComponent: function() {
-      var scope = context.getScope();
+    getParentStatelessComponent: function () {
+      let scope = context.getScope();
       while (scope) {
-        var node = scope.block;
-        var isFunction = /Function/.test(node.type); // Ignore non functions
-        var isNotMethod = !node.parent || node.parent.type !== 'MethodDefinition'; // Ignore classes methods
-        var isNotArgument = !node.parent || node.parent.type !== 'CallExpression'; // Ignore arguments (callback, etc.)
+        const node = scope.block;
+        // Ignore non functions
+        const isFunction = /Function/.test(node.type);
+        // Ignore classes methods
+        const isNotMethod = !node.parent || node.parent.type !== 'MethodDefinition';
+        // Ignore arguments (callback, etc.)
+        const isNotArgument = !node.parent || node.parent.type !== 'CallExpression';
         if (isFunction && isNotMethod && isNotArgument) {
           return node;
         }
@@ -241,31 +244,32 @@ function componentRule(rule, context) {
      * @param {ASTNode} node The AST node being checked (must be a MemberExpression).
      * @returns {ASTNode} component node, null if we cannot find the component
      */
-    getRelatedComponent: function(node) {
-      var i;
-      var j;
-      var k;
-      var l;
+    getRelatedComponent: function (node) {
+      let currentNode = node;
+      let i;
+      let j;
+      let k;
+      let l;
       // Get the component path
-      var componentPath = [];
-      while (node) {
-        if (node.property && node.property.type === 'Identifier') {
-          componentPath.push(node.property.name);
+      const componentPath = [];
+      while (currentNode) {
+        if (currentNode.property && currentNode.property.type === 'Identifier') {
+          componentPath.push(currentNode.property.name);
         }
-        if (node.object && node.object.type === 'Identifier') {
-          componentPath.push(node.object.name);
+        if (currentNode.object && currentNode.object.type === 'Identifier') {
+          componentPath.push(currentNode.object.name);
         }
-        node = node.object;
+        currentNode = currentNode.object;
       }
       componentPath.reverse();
 
       // Find the variable in the current scope
-      var variableName = componentPath.shift();
+      const variableName = componentPath.shift();
       if (!variableName) {
         return null;
       }
-      var variableInScope;
-      var variables = context.getScope().variables;
+      let variableInScope;
+      const variables = context.getScope().variables;
       for (i = 0, j = variables.length; i < j; i++) {
         if (variables[i].name === variableName) {
           variableInScope = variables[i];
@@ -277,10 +281,14 @@ function componentRule(rule, context) {
       }
 
       // Find the variable declaration
-      var defInScope;
-      var defs = variableInScope.defs;
+      let defInScope;
+      const defs = variableInScope.defs;
       for (i = 0, j = defs.length; i < j; i++) {
-        if (defs[i].type === 'ClassName' || defs[i].type === 'FunctionName' || defs[i].type === 'Variable') {
+        if (
+          defs[i].type === 'ClassName' ||
+          defs[i].type === 'FunctionName' ||
+          defs[i].type === 'Variable'
+        ) {
           defInScope = defs[i];
           break;
         }
@@ -288,72 +296,72 @@ function componentRule(rule, context) {
       if (!defInScope) {
         return null;
       }
-      node = defInScope.node.init || defInScope.node;
+      currentNode = defInScope.node.init || defInScope.node;
 
       // Traverse the node properties to the component declaration
       for (i = 0, j = componentPath.length; i < j; i++) {
-        if (!node.properties) {
+        if (!currentNode.properties) {
           continue;
         }
-        for (k = 0, l = node.properties.length; k < l; k++) {
-          if (node.properties[k].key.name === componentPath[i]) {
-            node = node.properties[k];
+        for (k = 0, l = currentNode.properties.length; k < l; k++) {
+          if (currentNode.properties[k].key.name === componentPath[i]) {
+            currentNode = currentNode.properties[k];
             break;
           }
         }
-        if (!node) {
+        if (!currentNode) {
           return null;
         }
-        node = node.value;
+        currentNode = currentNode.value;
       }
 
       // Return the component
-      return components.get(node);
-    }
+      return components.get(currentNode);
+    },
   };
 
   // Component detection instructions
-  var detectionInstructions = {
-    ClassDeclaration: function(node) {
+  const detectionInstructions = {
+    ClassDeclaration: function (node) {
       if (!utils.isES6Component(node)) {
         return;
       }
       components.add(node, 2);
     },
 
-    ClassProperty: function(node) {
-      node = utils.getParentComponent();
+    ClassProperty: function () {
+      const node = utils.getParentComponent();
       if (!node) {
         return;
       }
       components.add(node, 2);
     },
 
-    ObjectExpression: function(node) {
+    ObjectExpression: function (node) {
       if (!utils.isES5Component(node)) {
         return;
       }
       components.add(node, 2);
     },
 
-    FunctionExpression: function(node) {
-      node = utils.getParentComponent();
+    FunctionExpression: function () {
+      const node = utils.getParentComponent();
       if (!node) {
         return;
       }
       components.add(node, 1);
     },
 
-    FunctionDeclaration: function(node) {
-      node = utils.getParentComponent();
+    FunctionDeclaration: function () {
+      const node = utils.getParentComponent();
       if (!node) {
         return;
       }
       components.add(node, 1);
     },
 
-    ArrowFunctionExpression: function(node) {
-      node = utils.getParentComponent();
+    ArrowFunctionExpression: function () {
+      const node = utils.getParentComponent();
       if (!node) {
         return;
       }
@@ -364,8 +372,8 @@ function componentRule(rule, context) {
       }
     },
 
-    ThisExpression: function(node) {
-      node = utils.getParentComponent();
+    ThisExpression: function () {
+      const node = utils.getParentComponent();
       if (!node || !/Function/.test(node.type)) {
         return;
       }
@@ -373,23 +381,23 @@ function componentRule(rule, context) {
       components.add(node, 0);
     },
 
-    ReturnStatement: function(node) {
+    ReturnStatement: function (node) {
       if (!utils.isReturningJSX(node)) {
         return;
       }
-      node = utils.getParentComponent();
-      if (!node) {
+      const parentNode = utils.getParentComponent();
+      if (!parentNode) {
         return;
       }
-      components.add(node, 2);
-    }
+      components.add(parentNode, 2);
+    },
   };
 
   // Update the provided rule instructions to add the component detection
-  var ruleInstructions = rule(context, components, utils);
-  var updatedRuleInstructions = util._extend({}, ruleInstructions);
-  Object.keys(detectionInstructions).forEach(function(instruction) {
-    updatedRuleInstructions[instruction] = function(node) {
+  const ruleInstructions = rule(context, components, utils);
+  const updatedRuleInstructions = util._extend({}, ruleInstructions);
+  Object.keys(detectionInstructions).forEach((instruction) => {
+    updatedRuleInstructions[instruction] = (node) => {
       detectionInstructions[instruction](node);
       return ruleInstructions[instruction] ? ruleInstructions[instruction](node) : void 0;
     };
@@ -398,7 +406,7 @@ function componentRule(rule, context) {
   return updatedRuleInstructions;
 }
 
-Components.detect = function(rule) {
+Components.detect = function (rule) {
   return componentRule.bind(this, rule);
 };
 
