@@ -11,11 +11,11 @@
  * @returns {Boolean} True if the variable was found and marked as used, false if not.
  */
 function markVariableAsUsed(context, name) {
-  var scope = context.getScope();
-  var variables;
-  var i;
-  var len;
-  var found = false;
+  let scope = context.getScope();
+  let variables;
+  let i;
+  let len;
+  let found = false;
 
   // Special Node.js scope means we need to start one level deeper
   if (scope.type === 'global') {
@@ -45,8 +45,8 @@ function markVariableAsUsed(context, name) {
  * @returns {Boolean} True if the variable was found, false if not.
  */
 function findVariable(variables, name) {
-  var i;
-  var len;
+  let i;
+  let len;
 
   for (i = 0, len = variables.length; i < len; i++) {
     if (variables[i].name === name) {
@@ -67,8 +67,8 @@ function findVariable(variables, name) {
  * @returns {Boolean} True if the variable was found, false if not.
  */
 function variablesInScope(context) {
-  var scope = context.getScope();
-  var variables = scope.variables;
+  let scope = context.getScope();
+  let variables = scope.variables;
 
   while (scope.type !== 'global') {
     scope = scope.upper;
@@ -87,5 +87,5 @@ function variablesInScope(context) {
 module.exports = {
   findVariable: findVariable,
   variablesInScope: variablesInScope,
-  markVariableAsUsed: markVariableAsUsed
+  markVariableAsUsed: markVariableAsUsed,
 };
